Added initial UI Event classes
