+++ src/vsg/core/ConstVisitor.cpp
@@ -22,6 +22,9 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 #include <vsg/vk/CommandBuffer.h>
 #include <vsg/vk/RenderPass.h>
 
+#include <vsg/ui/KeyEvent.h>
+#include <vsg/ui/PointerEvent.h>
+
 using namespace vsg;
 
 ConstVisitor::ConstVisitor()
@@ -292,4 +295,37 @@ void ConstVisitor::apply(const CommandBuffer& value)
 void ConstVisitor::apply(const RenderPass& value)
 {
     apply(static_cast<const Object&>(value));
+}
+
+////////////////////////////////////////////////////////////////////////////////
+//
+// UI Events
+//
+void ConstVisitor::apply(const UIEvent& event)
+{
+    apply(static_cast<const Object&>(event));
+}
+void ConstVisitor::apply(const KeyEvent& event)
+{
+    apply(static_cast<const UIEvent&>(event));
+}
+void ConstVisitor::apply(const KeyPressEvent& event)
+{
+    apply(static_cast<const KeyEvent&>(event));
+}
+void ConstVisitor::apply(const KeyReleaseEvent& event)
+{
+    apply(static_cast<const KeyEvent&>(event));
+}
+void ConstVisitor::apply(const PointerEvent& event)
+{
+    apply(static_cast<const UIEvent&>(event));
+}
+void ConstVisitor::apply(const ButtonPressEvent& event)
+{
+    apply(static_cast<const PointerEvent&>(event));
+}
+void ConstVisitor::apply(const ButtonReleaseEvent& event)
+{
+    apply(static_cast<const PointerEvent&>(event));
 }
